+++ R/smocc_200.R
@@ -2,17 +2,19 @@
 #'
 #' Longitudinal height and weight measurements during ages 0-2 years for a
 #' representative sample of 1933 Dutch children born in 1988-1989. The dataset
-#' \code{smocc_200} is a subset of the full data covering 200 children.
+#' \code{smocc_200} is sample of size 200 from the full data.
 #'
 #' @name smocc_200
 #' @docType data
-#' @format A tibble with 1940 rows and 7 columns: \describe{ \item{id}{ID,
-#' unique \code{id} of each child (numeric)} \item{age}{Decimal age, 0-2.12
-#' years (numeric)} \item{sex}{Sex, \code{"male"} or \code{"female"}
-#' (character)} \item{ga}{Gestational age, completed weeks (numeric)}
-#' \item{bw}{Birth weight in grammes (numeric)} \item{hgt}{Height measurement in
-#' cm (34-102) (numeric)} \item{hgt_z}{Height in SDS relative Fourth Dutch
-#' Growth Study 1997 (numeric)} }
+#' @format A tibble with 1942 rows and 7 columns:
+#' \describe{
+#' \item{id}{ID, unique \code{id} of each child (numeric)}
+#' \item{age}{Decimal age, 0-2.68 years (numeric)}
+#' \item{sex}{Sex, \code{"male"} or \code{"female"} (character)}
+#' \item{ga}{Gestational age, completed weeks (numeric)}
+#' \item{bw}{Birth weight in grammes (numeric)}
+#' \item{hgt}{Height measurement in cm (numeric)}
+#' \item{hgt_z}{Height in SDS relative Fourth Dutch Growth Study 1997 (numeric)}}
 #' @source Herngreen WP, van Buuren S, van Wieringen JC, Reerink JD,
 #'   Verloove-Vanhorick SP & Ruys JH (1994). Growth in length and weight from
 #'   birth to 2 years of a representative sample of Netherlands children (born
@@ -37,7 +39,7 @@ NULL
 #' ```{r eval=FALSE}
 #' knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24)/12, 4)
 #' fit_200 <- brokenstick(hgt_z ~ age | id, data = smocc_200,
-#'                        knots = knots, boundary = c(0, 3), seed = 1)
+#'                        knots = knots, seed = 1)
 #' ```
 #' @keywords datasets
 NULL
@@ -57,7 +59,7 @@ NULL
 #' ```{r eval=FALSE}
 #' knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24)/12, 4)
 #' fit_200_light <- brokenstick(hgt_z ~ age | id, data = smocc_200,
-#'                        knots = knots, boundary = c(0, 3),
+#'                        knots = knots,
 #'                        light = TRUE, seed = 1)
 #' ```
 #' @keywords datasets