+++ data-raw/R/smocc_200.R
@@ -24,16 +24,11 @@ hgt_z <- with(
 smocc_200$hgt_z <- hgt_z
 
 # fit the brokenstick model
-knots <- round(c(0, 1, 2, 3, 6, 9, 12, 15, 18, 24) / 12, 4)
-fit_200 <- brokenstick(hgt_z ~ age | id,
-  data = smocc_200,
-  knots = knots, boundary = c(0, 3), seed = 1
-)
-fit_200_light <- brokenstick(hgt_z ~ age | id,
-  data = smocc_200,
-  knots = knots, boundary = c(0, 3),
-  light = TRUE, seed = 1
-)
+fit_200 <- brokenstick(hgt_z ~ age | id, smocc_200, seed = 1,
+                    knots = round(c(0:3, 6, 9, 12, 15, 18, 24)/12, 4))
+fit_200_light <- brokenstick(hgt_z ~ age | id, smocc_200, seed = 1,
+                    knots = round(c(0:3, 6, 9, 12, 15, 18, 24)/12, 4),
+                    light = TRUE)
 usethis::use_data(smocc_200, overwrite = TRUE)
 usethis::use_data(fit_200, overwrite = TRUE)
 usethis::use_data(fit_200_light, overwrite = TRUE)